LLAMA-6895 : Change kwd sensitivity to a floating point
Reason for change: Change keyword sensitivity value from detector
specific integer values and ranges to flexible and consistent floating
point values. Limit checking to be handled by the detector.
Test Procedure: Test and verify sensitivity values in configuration
files are applied correctly to applicable keyword detectors
Risks: Low

(cherry picked from commit a93ae8e2f99742ad36c707a5294bfca178f999b4)

+++ src/xrsr.h
@@ -459,7 +459,7 @@ typedef struct {
 /// @brief XRSR keyword configuration structure
 /// @details The keyword configuration data structure indicates detailed information for the keyword detector.
 typedef struct {
-   uint32_t sensitivity; ///< Sensitivity level of the keyword detector
+   float sensitivity; ///< Sensitivity level of the keyword detector
 } xrsr_keyword_config_t;
 
 /// @brief XRSR capture config structure
